+++ tests/test_webhook_payloads.py
@@ -19,8 +19,8 @@ def _remove_anonymized_order_data(order_data: dict) -> dict:
     del order_data[0]["user_email"]
     del order_data[0]["billing_address"]
     del order_data[0]["shipping_address"]
-    del order_data[0]["meta"]
-    del order_data[0]["private_meta"]
+    del order_data[0]["metadata"]
+    del order_data[0]["private_metadata"]
     return order_data
 
 
@@ -53,8 +53,8 @@ def test_generate_sample_payload_order(
         order.shipping_address.street_address_1
         != payload[0]["shipping_address"]["street_address_1"]
     )
-    assert order.meta != payload[0]["meta"]
-    assert order.private_meta != payload[0]["private_meta"]
+    assert order.metadata != payload[0]["metadata"]
+    assert order.private_metadata != payload[0]["private_metadata"]
     # Remove anonymized data
     payload = _remove_anonymized_order_data(payload)
     order_payload = _remove_anonymized_order_data(order_payload)
@@ -99,8 +99,8 @@ def _remove_anonymized_checkout_data(checkout_data: dict) -> dict:
     del checkout_data[0]["email"]
     del checkout_data[0]["billing_address"]
     del checkout_data[0]["shipping_address"]
-    del checkout_data[0]["meta"]
-    del checkout_data[0]["private_meta"]
+    del checkout_data[0]["metadata"]
+    del checkout_data[0]["private_metadata"]
     return checkout_data
 
 
@@ -121,8 +121,8 @@ def test_generate_sample_checkout_payload(user_checkout_with_items):
         != payload[0]["shipping_address"]["street_address_1"]
     )
     assert "note" not in payload[0]
-    assert checkout.meta != payload[0]["meta"]
-    assert checkout.private_meta != payload[0]["private_meta"]
+    assert checkout.metadata != payload[0]["metadata"]
+    assert checkout.private_metadata != payload[0]["private_metadata"]
     # Remove anonymized data
     payload = _remove_anonymized_checkout_data(payload)
     checkout_payload = _remove_anonymized_checkout_data(checkout_payload)