+++ CHANGELOG.md
@@ -66,6 +66,14 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix passing addresses to `PaymentData` objects - #5223 by @maarcingebala
 - Return `null` when querying `me` as an anonymous user - #5231 by @maarcingebala
 - Added `PLAYGROUND_ENABLED` environment variable/setting to allow to enable the GraphQL playground when `DEBUG` is disabled - #5254 by @NyanKiyoshi
+- Fix access to order query when request from service account - #5258 by @fowczarek
+- Customer shouldn't be able to see draft orders by token - #5259 by @fowczarek
+- Customer shouldn't be able to query checkout with another customer - #5268 by @fowczarek
+- Added integration support of Jaeger Tracing - #5282 by @NyanKiyoshi
+- Customer shouldn't be able to see draft orders by token  - #5259 by @fowczarek
+- Return `null` when querying `me` as an anonymous user - #5231 as @maarcingebala
+- Unify saleor metadata - #5178 by @fowczarek
+
 
 ## 2.9.0